+++ media/transcoder/openjpeg/j2klib/linux_amd64/libopenjpeg.a
