+++ sonar-project.properties
@@ -1,5 +1,5 @@
 sonar.coverage.exclusions=**/*_test.go
-sonar.exclusions=jpeglib/**/*,openjpeg/**/*,**/*_test.go
+sonar.exclusions=jpeglib/**/*,media/transcoder/**/*,**/*_test.go
 sonar.go.coverage.reportPaths=./coverage.out
 sonar.language=go
 sonar.organization=t2care
